# Use Daggers eager API with (vendored) lazy trees