--- a/test/basics.jl
+++ b/test/basics.jl
@@ -230,7 +230,8 @@ end
     end
 end
 
-@testset "exec with context" begin
+# TODO: Just remove if we anyways can't select which context to use in Daggers eager/standard API?
+#= @testset "exec with context" begin
     import Dagger
 
     struct SpecialContext end
@@ -251,7 +252,7 @@ end
     @test exec(SpecialContext(), t) |> values == [1,2,3]
     @test computespecial[] == true # Dummy compare to make failed test outprint a little less confusing
     @test ncollectspecial[] == 3 # All values are collected with SpecialContext
-end
+end =#
 
 
 @testset "iterators" begin