@DrChainsaw 2026-03-08
@shashi : I'm thinking of going ahead here, but I'd like you to just review at least the high level changes described below.

The big change is that I made Dagger a weak dependency (since we no longer need it for the lazyness). At least for me, in 99% of my use cases Dagger is overkill and I often see lower latency and memory usage with the simpler built in executors.

This will be a breaking release, but the breakage should not affect most users. I have added the reasons for the breakage in the changelog.

I have ran with these changes for a while in some of my analysis scripts, so there is a little bit more testing done than just the unit-tests.